+++ odbc-api/src/result_set_metadata.rs
@@ -182,6 +182,129 @@ pub trait ResultSetMetadata {
         };
         Ok(dt)
     }
+
+    /// The maximum binary length of UTF-8 representation of elements of this column. You may
+    /// specify the data type of the column to safe function calls into ODBC. Call this method to
+    /// determine maxmimum string length for UTF-8 text buffers.
+    ///
+    /// This method assumes the encoding of text columns to be UTF-8.
+    fn utf8_len(&self, column_number: u16, hint: Option<DataType>) -> Result<usize, Error> {
+        let stmt = self.stmt_ref();
+        let sql_type = if let Some(dt) = hint {
+            dt.data_type()
+        } else {
+            stmt.col_concise_type(column_number).into_result(stmt)?
+        };
+
+        if is_narrow_text(sql_type) {
+            // Query binary length
+            return Ok(self.col_octet_length(column_number)?.try_into().unwrap());
+        }
+
+        if is_wide_text(sql_type) {
+            // Utf16 to Utf8 conversion. We must adjust the binary size reported by the database,
+            // since it assumes UTF-16 encoding. While usually an UTF-8 string is shorter for
+            // characters in the range from U+0800 to U+FFFF UTF-8 is three bytes theras UTF-16 is
+            // only two bytes large. We have to allocate enough memory to hold the largest possible
+            // string.
+            let octet_len: usize = self.col_octet_length(column_number)?.try_into().unwrap();
+            // Utf16 to Utf8 conversion. We must adjust the binary size reported by the database,
+            // since it assumes UTF-16 encoding. While usually an UTF-8 string is shorter for
+            // characters in the range from U+0800 to U+FFFF UTF-8 is three bytes theras UTF-16 is
+            // only two bytes large. We have to allocate enough memory to hold the largest possible
+            // string.
+            return Ok((octet_len / 2) * 3)
+        }
+
+        // Assume non text columns to be representable in ASCII. Look if we can display size
+        // from hint, otherwise ask the driver.
+        let binary_len = if let Some(display_size) = sql_type_to_display_size(sql_type) {
+            display_size
+        } else if let Some(display_size) = hint.and_then(|dt| dt.display_size()) {
+            display_size
+        } else {
+            self.col_display_size(column_number)?.try_into().unwrap()
+        };
+        Ok(binary_len)
+    }
+
+    /// The maximum length of the UTF-16 representation in 2-Byte characters.
+    ///
+    /// You may specify the data type of the column to safe function calls into ODBC. Call this
+    /// method to determine maxmimum string length for UTF-8 text buffers.
+    fn utf16_len(&self, column_number: u16, hint: Option<DataType>) -> Result<usize, Error> {
+        let stmt = self.stmt_ref();
+        let sql_type = if let Some(dt) = hint {
+            dt.data_type()
+        } else {
+            stmt.col_concise_type(column_number).into_result(stmt)?
+        };
+
+        if is_narrow_text(sql_type) {
+            // Utf8 to Utf16 conversion. The letters in the range from U+0000 to U+007F take two
+            // bytes in UTF-16 but only one byte in UTF-8. Since we return the length in `u16` the
+            // value is divided and multiplied by two => the UTF-8 octet length is the UTF-16 len.
+            return Ok(self.col_octet_length(column_number)?.try_into().unwrap());
+        }
+
+        if is_wide_text(sql_type) {
+            let octet_len: usize = self.col_octet_length(column_number)?.try_into().unwrap();
+            // Octet is the same as a byte. A `u16` consists of two bytes.
+            return Ok(octet_len / 2)
+        }
+
+        // Assume non text columns to be representable in ASCII. Look if we can display size
+        // from hint, otherwise ask the driver.
+        let binary_len = if let Some(display_size) = sql_type_to_display_size(sql_type) {
+            display_size
+        } else if let Some(display_size) = hint.and_then(|dt| dt.display_size()) {
+            display_size
+        } else {
+            self.col_display_size(column_number)?.try_into().unwrap()
+        };
+        Ok(binary_len)
+    }
+}
+
+/// Should the information of the SQL type alone, be enough to determine the maximum display size
+/// this function will return it.
+fn sql_type_to_display_size(sql_type: SqlDataType) -> Option<usize> {
+    let ret = match sql_type {
+        // 11 if signed (a sign and 10 digits) or 10 if unsigned (10 digits).
+        SqlDataType::INTEGER => 11,
+        // 6 if signed (a sign and 5 digits) or 5 if unsigned (5 digits).
+        SqlDataType::SMALLINT => 6,
+        // 24 (a sign, 15 digits, a decimal point, the letter E, a sign, and 3 digits).
+        SqlDataType::DOUBLE => 24,
+        // 14 (a sign, 7 digits, a decimal point, the letter E, a sign, and 2 digits).
+        SqlDataType::REAL => 14,
+        // 10 (a date in the format yyyy-mm-dd).
+        SqlDataType::DATE => 10,
+        // 20 (a sign and 19 digits if signed or 20 digits if unsigned).
+        SqlDataType::EXT_BIG_INT => 20,
+        // 4 if signed (a sign and 3 digits) or 3 if unsigned (3 digits).
+        SqlDataType::EXT_TINY_INT => 4,
+        // 1 digit.
+        SqlDataType::EXT_BIT => 1,
+        _ => return None,
+    };
+    Some(ret)
+}
+
+/// True if the SQL data type is narrow text (usually utf8) on the DBC
+fn is_narrow_text(sql_type: SqlDataType) -> bool {
+    matches!(
+        sql_type,
+        SqlDataType::CHAR | SqlDataType::VARCHAR | SqlDataType::EXT_LONG_VARCHAR
+    )
+}
+
+/// True if the SQL data type is wide text (utf16) on the DBC
+fn is_wide_text(sql_type: SqlDataType) -> bool {
+    matches!(
+        sql_type,
+        SqlDataType::EXT_W_CHAR | SqlDataType::EXT_W_VARCHAR | SqlDataType::EXT_W_LONG_VARCHAR
+    )
 }
 
 /// An iterator calling `col_name` for each column_name and converting the result into UTF-8. See